Move abstract methods from `BaseWorker` to `AbstractWorker`
